Add missing commas so that "use strict" works (#161)
Co-Authored-By: jpaddison3 <jp@centreforeffectivealtruism.org>

When we experimented with adding "use strict" to our codebase in a way that accidentally covered all of `node_modules`, this file crashed on start because `crypto` is an assignment to an undeclared variable. It's an undeclared variable because the preceding line is missing a comma. Fix that, so that this file can pass `use strict`.

Co-authored-by: Frederik Prijck <frederik.prijck@auth0.com>